Moves Breadcrumbs directory to Builder directory
